Match-id-85119607e205823b9d3a468d46a9694cfcf3dd61

+++ include/msServiceProfiler/SecurityUtils.h
@@ -17,7 +17,7 @@ namespace SecurityUtils {
     bool IsDir(std::string const &absPath); // 检查是否为文件夹
     bool IsPathLenLegal(const std::string &absPath); // 检查路径长度是否合法
     bool IsPathDepthLegal(const std::string &absPath); // 检查路径深度是否合法
-    bool IsFileSizeLegal(const std::string &absPath, long long maxSize = MAX_FILE_SIZE_10G); // 检查文件大小是否合法
+    bool IsFileSizeLegal(const std::string &absPath, uint64_t maxSize = MAX_FILE_SIZE_10G); // 检查文件大小是否合法
     bool IsPathCharactersValid(const std::string &absPath); // 检查路径中是否有非法字符
 
     bool CheckPathContainSoftLink(const std::string &path); // 检查路径中各层级是否为软链接
